+++ paymentsheet/src/main/java/com/stripe/android/customersheet/CustomerSheet.kt
@@ -3,7 +3,6 @@ package com.stripe.android.customersheet
 import android.app.Application
 import androidx.activity.ComponentActivity
 import androidx.activity.result.ActivityResultRegistryOwner
-import androidx.annotation.RestrictTo
 import androidx.core.app.ActivityOptionsCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.DefaultLifecycleObserver
@@ -15,16 +14,17 @@ import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.paymentsheet.model.PaymentOptionFactory
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.utils.AnimationConstants
+import kotlinx.coroutines.async
+import kotlinx.coroutines.coroutineScope
 import javax.inject.Inject
 
 /**
- * 🏗 This feature is under construction 🏗
+ * 🏗 This feature is in private beta and could change 🏗
  *
  * [CustomerSheet] A class that presents a bottom sheet to manage a customer through the
  * [CustomerAdapter].
  */
 @ExperimentalCustomerSheetApi
-@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 class CustomerSheet @Inject internal constructor(
     private val application: Application,
     lifecycleOwner: LifecycleOwner,
@@ -81,10 +81,16 @@ class CustomerSheet @Inject internal constructor(
      * Retrieves the customer's saved payment option selection or null if the customer does not have
      * a persisted payment option selection.
      */
-    suspend fun retrievePaymentOptionSelection(): CustomerSheetResult {
+    suspend fun retrievePaymentOptionSelection(): CustomerSheetResult = coroutineScope {
         val adapter = CustomerSessionViewModel.component.customerAdapter
-        val selectedPaymentOption = adapter.retrieveSelectedPaymentOption()
-        val paymentMethods = adapter.retrievePaymentMethods()
+        val selectedPaymentOptionDeferred = async {
+            adapter.retrieveSelectedPaymentOption()
+        }
+        val paymentMethodsDeferred = async {
+            adapter.retrievePaymentMethods()
+        }
+        val selectedPaymentOption = selectedPaymentOptionDeferred.await()
+        val paymentMethods = paymentMethodsDeferred.await()
 
         val selection = selectedPaymentOption.mapCatching { paymentOption ->
             paymentOption?.toPaymentSelection {
@@ -94,26 +100,25 @@ class CustomerSheet @Inject internal constructor(
             }?.toPaymentOptionSelection(paymentOptionFactory)
         }
 
-        return selection.fold(
+        selection.fold(
             onSuccess = {
                 CustomerSheetResult.Selected(it)
             },
             onFailure = { cause, _ ->
-                CustomerSheetResult.Error(cause)
+                CustomerSheetResult.Failed(cause)
             }
         )
     }
 
     private fun onCustomerSheetResult(result: InternalCustomerSheetResult?) {
         requireNotNull(result)
-        callback.onResult(result.toPublicResult(paymentOptionFactory))
+        callback.onCustomerSheetResult(result.toPublicResult(paymentOptionFactory))
     }
 
     /**
      * Configuration for [CustomerSheet]
      */
     @ExperimentalCustomerSheetApi
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     class Configuration internal constructor(
         /**
          * Describes the appearance of [CustomerSheet].
@@ -154,6 +159,16 @@ class CustomerSheet @Inject internal constructor(
         val merchantDisplayName: String? = null,
     ) {
 
+        // Hide no-argument constructor init
+        internal constructor() : this(
+            appearance = PaymentSheet.Appearance(),
+            googlePayEnabled = false,
+            headerTextForSelectionScreen = null,
+            defaultBillingDetails = PaymentSheet.BillingDetails(),
+            billingDetailsCollectionConfiguration = PaymentSheet.BillingDetailsCollectionConfiguration(),
+            merchantDisplayName = null,
+        )
+
         fun newBuilder(): Builder {
             return Builder()
                 .appearance(appearance)
@@ -165,8 +180,7 @@ class CustomerSheet @Inject internal constructor(
         }
 
         @ExperimentalCustomerSheetApi
-        @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-        class Builder {
+        class Builder internal constructor() {
             private var appearance: PaymentSheet.Appearance = PaymentSheet.Appearance()
             private var googlePayEnabled: Boolean = false
             private var headerTextForSelectionScreen: String? = null
@@ -211,10 +225,17 @@ class CustomerSheet @Inject internal constructor(
                 merchantDisplayName = merchantDisplayName,
             )
         }
+
+        companion object {
+
+            @JvmStatic
+            fun builder(): Builder {
+                return Builder()
+            }
+        }
     }
 
     @ExperimentalCustomerSheetApi
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     companion object {
 
         /**
@@ -225,6 +246,7 @@ class CustomerSheet @Inject internal constructor(
          * @param customerAdapter The bridge to communicate with your server to manage a customer.
          * @param callback called when a [CustomerSheetResult] is available.
          */
+        @JvmStatic
         fun create(
             activity: ComponentActivity,
             configuration: Configuration,
@@ -250,6 +272,7 @@ class CustomerSheet @Inject internal constructor(
          * @param customerAdapter The bridge to communicate with your server to manage a customer.
          * @param callback called when a [CustomerSheetResult] is available.
          */
+        @JvmStatic
         fun create(
             fragment: Fragment,
             configuration: Configuration,